remove useLoadStore, add useAnalytics hook
