+++ src/routes/safe/container/index.jsx
@@ -1,11 +1,18 @@
 // @flow
 import * as React from 'react'
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
 
 import Page from '~/components/layout/Page'
+import loadAddressBookFromStorage from '~/logic/addressBook/store/actions/loadAddressBookFromStorage'
+import addViewedSafe from '~/logic/currentSession/store/actions/addViewedSafe'
 import { type Token } from '~/logic/tokens/store/model/token'
 import Layout from '~/routes/safe/components/Layout'
 import { useCheckForUpdates } from '~/routes/safe/container/Hooks/useCheckForUpdates'
+import fetchLatestMasterContractVersion from '~/routes/safe/store/actions/fetchLatestMasterContractVersion'
+import fetchSafe from '~/routes/safe/store/actions/fetchSafe'
+import fetchTransactions from '~/routes/safe/store/actions/fetchTransactions'
+import { safeParamAddressFromStateSelector } from '~/routes/safe/store/selectors'
 
 type Action = 'Send' | 'Receive'
 
@@ -17,12 +24,29 @@ const INITIAL_STATE = {
   showReceive: false,
 }
 
-type Props = {
-  safeLoaded: boolean,
-}
-
-const SafeView = (props: Props) => {
+const SafeView = () => {
   const [state, setState] = useState(INITIAL_STATE)
+  const [safeLoaded, setSafeLoaded] = useState(false)
+  const dispatch = useDispatch()
+
+  const safeAddress = useSelector(safeParamAddressFromStateSelector)
+
+  useEffect(() => {
+    const fetchData = () => {
+      if (safeAddress) {
+        dispatch(fetchLatestMasterContractVersion())
+          .then(() => dispatch(fetchSafe(safeAddress)))
+          .then(() => {
+            setSafeLoaded(true)
+            dispatch(loadAddressBookFromStorage())
+            return dispatch(fetchTransactions(safeAddress))
+          })
+          .then(() => dispatch(addViewedSafe(safeAddress)))
+      }
+    }
+    fetchData()
+  }, [safeAddress])
+
   useCheckForUpdates()
 
   const onShow = (action: Action) => () => {
@@ -62,7 +86,7 @@ const SafeView = (props: Props) => {
 
   return (
     <Page>
-      {!props.safeLoaded ? null : (
+      {!safeLoaded ? null : (
         <>
           <Layout
             hideSendFunds={hideSendFunds}
@@ -78,4 +102,4 @@ const SafeView = (props: Props) => {
   )
 }
 
-export default SafeView
+export default React.memo(SafeView)
